Serialization problem fixed
Pickling to binary and JSON works now.
JSON: Originally, the variable dict_from_csv was assigned outside the functions for the dictionary creation and serialization. This resulted in the dictionary dict_from_csv being blank when creating a json file.
Instead now, I have declared the dict_from_csv variable to be global within the dictionary creating function, allowing it to be manipulated by other functions.

BINARY:  I was writing the output of the pickle.dump into a string format instead of the binary format.  In line 144 I changed "w" into "wb".